Begin adding in stuff for updating app, downloads are returning (permission denied) even with the sd verified to be writeable and ready to go. SU was attempted and gained, but still permission denied.
Included is a partial logcat shoing the sd state, the SU request, and then the permission denied error.
I feel like this is the same error holding back the other downloads, and could also be to blame for the java IO methods of creating and deleting the directories not working.