Launch the market app
Aloow launching of the market app. Warn user if no activity found
to luanch the market, ie no market/gapps installed.
If the file does exist for flashing and the packages is
not a market app it should determine weather to flash or not
depeding if the file is finished.(Performs a byte check)